requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dicttoxml==1.7.4
flake8==3.8.1
flasgger==0.9.4
Flask==1.1.2
Flask-Cors==3.0.8
Flask-HTTPAuth==4.0.0
flask-marshmallow==0.12.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
json2xml==3.3.2
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-sqlalchemy==0.23.0
mccabe==0.6.1
mistune==0.8.4
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
requests==2.23.0
schematics==2.1.0
six==1.11.0
SQLAlchemy==1.3.16
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
wikitextparser==0.36.1
xmltodict==0.11.0
zipp==3.1.0
